+++ src/lib/puzzles.test.ts
@@ -1,5 +1,6 @@
 import { expect, it } from 'vitest';
 import { puzzles } from '$lib/puzzles';
+import { isPuzzleSolvable } from './math';
 
 it('Unqiue puzzle titles', () => {
 	const uniqueValues = new Set();
@@ -14,4 +15,19 @@ it('Unqiue puzzle titles', () => {
 	}
 
 	expect(duplicates).toHaveLength(0);
+});
+
+it('Solvable puzzles', () => {
+	const solvable = new Set();
+	const unsolvable = [];
+
+	for (const { title, start, end } of puzzles) {
+		if (isPuzzleSolvable(start, end)) {
+			solvable.add(title);
+		} else {
+			unsolvable.push(title);
+		}
+	}
+
+	expect(unsolvable).toHaveLength(0);
 });